REFACTOR: refactor Package Name fcm to firebase
