+++ voicepocket/src/main/java/com/vp/voicepocket/domain/firebase/repository/FCMRepository.java
@@ -1,6 +1,6 @@
-package com.vp.voicepocket.domain.fcm.repository;
+package com.vp.voicepocket.domain.firebase.repository;
 
-import com.vp.voicepocket.domain.fcm.entity.FCMUserToken;
+import com.vp.voicepocket.domain.firebase.entity.FCMUserToken;
 import com.vp.voicepocket.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
